+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-09-08
+
+### Added
+
+- "(re)start" button
+- index of the current task in the middle of the header
+
+### Changed
+
+- "session" length is 10 tasks
+
 ## [0.1.0] - 2023-08-24
 
 ### Added
@@ -15,6 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - autofocus on input after page is loaded
 - score with right answers and total count of problems
 
-[Unreleased]: https://github.com/vikian050194/click/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/vikian050194/click/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/vikian050194/click/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/vikian050194/click/releases/tag/v0.1.0